requirements.txt
regex
streamlit
pandas
wordcloud
emoji
matplotlib
seaborn
plotly
Pillow
numpy
emot
nltk
streamlit-aggrid
